diet(phase8): split tools/session/project_ops.rs (1399 LOC) into 8-file module
Mechanical split, behavior unchanged. session/mod.rs:17 `pub use project_ops::{...}`
resolves the same 3 symbols via the new mod.rs re-exports.

tools/session/project_ops/
- mod.rs (13)            re-exports
- activate.rs (123)      activate_project
- prepare_harness.rs (625)  prepare_harness_session (the main fn)
- prep_warnings.rs (199) push/append/collect warning helpers
- prep_recovery.rs (158) RefreshSymbolIndexRemediation +
                         refresh_*_for_surface +
                         prepare_harness_index_recovery + index_stats_payload
- util.rs (34)           is_anonymized_agent_project_name,
                         client_tool_schema_fingerprint
- embed_hint.rs (60)     auto_set_embed_hint_lang
- tests.rs (233)         mod tests content

Gates:
- cargo check --workspace --features http
- cargo test -p codelens-mcp --bin codelens-mcp (593 passed, 0 failed)
- cargo fmt --all -- --check
