--- crates/codelens-mcp/src/tools/session/project_ops.rs
@@ -1,1399 +1,0 @@
-use crate::AppState;
-use crate::protocol::BackendKind;
-use crate::resource_context::{
-    ResourceRequestContext, build_http_session_payload, build_visible_tool_context,
-};
-use crate::tool_defs::{
-    HostContext, TaskOverlay, ToolPreset, ToolProfile, ToolSurface, compile_surface_overlay,
-    default_budget_for_profile, is_tool_in_surface, preferred_bootstrap_tools, tool_name_requests,
-    tool_request_omissions,
-};
-use crate::tool_runtime::{ToolResult, success_meta};
-use codelens_engine::memory::list_memory_names;
-use codelens_engine::{compute_dominant_language, detect_frameworks};
-use serde_json::{Value, json};
-use std::collections::HashSet;
-
-const DEFAULT_AUTO_REFRESH_STALE_THRESHOLD: usize = 32;
-
-/// Issue #186: heuristic for spotting Claude/Codex internal workspace
-/// directories whose basename is itself an anonymized agent hash
-/// (e.g. `agent-a110134bd9c6e7440`). When `prepare_harness_session`
-/// resolves the active project to one of these instead of the
-/// daemon's CLI startup root, downstream tools see a near-empty
-/// index and emit false-positive `no_supported_files` warnings.
-///
-/// We treat any name that begins with `agent-` and whose suffix is
-/// strictly hex (≥ 12 chars of `[0-9a-fA-F]`) as anonymized — the
-/// shape that the harness itself produces. Limiting to hex avoids
-/// false-positives on real project directories like `agent-server/`
-/// or `agent-orchestrator/` where alphabetic chars like `r`/`s`/`t`
-/// rule out a hash interpretation.
-fn is_anonymized_agent_project_name(name: &str) -> bool {
-    let Some(suffix) = name.strip_prefix("agent-") else {
-        return false;
-    };
-    suffix.len() >= 12 && suffix.chars().all(|c| c.is_ascii_hexdigit())
-}
-
-fn push_prepare_harness_warning(
-    warnings: &mut Vec<Value>,
-    warning_codes: &mut HashSet<String>,
-    code: &str,
-    message: &str,
-    restart_recommended: bool,
-    recommended_action: &str,
-    action_target: &str,
-) {
-    if warning_codes.insert(code.to_owned()) {
-        warnings.push(json!({
-            "code": code,
-            "message": message,
-            "restart_recommended": restart_recommended,
-            "recommended_action": recommended_action,
-            "action_target": action_target,
-        }));
-    }
-}
-
-/// Variant of `push_prepare_harness_warning` that merges `extras` into
-/// the warning object, used when a warning carries actionable
-/// follow-up data (e.g. `remediation`, `auto_refresh_threshold`,
-/// per-file breakdown). Existing callers can stay on the simpler
-/// helper; only warnings that need to surface concrete next steps
-/// or freshness breakdowns reach for this variant.
-///
-/// `extras` is expected to be a `Value::Object`. Non-object values are
-/// ignored so the warning shape stays consistent.
-#[allow(clippy::too_many_arguments)]
-fn push_prepare_harness_warning_with_extras(
-    warnings: &mut Vec<Value>,
-    warning_codes: &mut HashSet<String>,
-    code: &str,
-    message: &str,
-    restart_recommended: bool,
-    recommended_action: &str,
-    action_target: &str,
-    extras: Value,
-) {
-    if !warning_codes.insert(code.to_owned()) {
-        return;
-    }
-    let mut warning = serde_json::Map::new();
-    warning.insert("code".to_owned(), json!(code));
-    warning.insert("message".to_owned(), json!(message));
-    warning.insert("restart_recommended".to_owned(), json!(restart_recommended));
-    warning.insert("recommended_action".to_owned(), json!(recommended_action));
-    warning.insert("action_target".to_owned(), json!(action_target));
-    if let Value::Object(map) = extras {
-        for (key, value) in map {
-            warning.insert(key, value);
-        }
-    }
-    warnings.push(Value::Object(warning));
-}
-
-fn client_tool_schema_fingerprint(arguments: &Value) -> Option<&str> {
-    [
-        "known_tool_schema_fingerprint",
-        "client_tool_schema_fingerprint",
-        "_session_tool_schema_fingerprint",
-        "_tool_schema_fingerprint",
-    ]
-    .iter()
-    .find_map(|key| arguments.get(*key).and_then(Value::as_str))
-    .map(str::trim)
-    .filter(|value| !value.is_empty())
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum RefreshSymbolIndexRemediation {
-    Force,
-    StaleOnly,
-}
-
-impl RefreshSymbolIndexRemediation {
-    fn command(self) -> &'static str {
-        match self {
-            Self::Force => "codelens reindex --force",
-            Self::StaleOnly => "codelens reindex --stale-only",
-        }
-    }
-
-    fn args(self) -> Option<Value> {
-        match self {
-            Self::Force => None,
-            Self::StaleOnly => Some(json!({ "scope": "stale_only" })),
-        }
-    }
-}
-
-fn refresh_symbol_index_remediation_for_surface(
-    surface: ToolSurface,
-    mode: RefreshSymbolIndexRemediation,
-) -> Value {
-    let command = mode.command();
-    let args = mode.args();
-    let tool_callable = is_tool_in_surface("refresh_symbol_index", surface);
-
-    if tool_callable {
-        let mut remediation = serde_json::Map::new();
-        remediation.insert("method".to_owned(), json!("tool_call"));
-        remediation.insert("tool".to_owned(), json!("refresh_symbol_index"));
-        if let Some(args) = args {
-            remediation.insert("args".to_owned(), args);
-        }
-        remediation.insert("callable".to_owned(), json!(true));
-        remediation.insert("alternative_command".to_owned(), json!(command));
-        return Value::Object(remediation);
-    }
-
-    let mut tool_call = serde_json::Map::new();
-    tool_call.insert("tool".to_owned(), json!("refresh_symbol_index"));
-    if let Some(args) = args {
-        tool_call.insert("args".to_owned(), args);
-    }
-    tool_call.insert("callable".to_owned(), json!(false));
-    tool_call.insert("reason".to_owned(), json!("not_in_active_surface"));
-    tool_call.insert("surface".to_owned(), json!(surface.as_label()));
-
-    json!({
-        "method": "shell",
-        "command": command,
-        "alternative_command": command,
-        "tool_call": Value::Object(tool_call),
-    })
-}
-
-fn refresh_symbol_index_recommended_action_for_surface(surface: ToolSurface) -> &'static str {
-    if is_tool_in_surface("refresh_symbol_index", surface) {
-        "refresh_symbol_index"
-    } else {
-        "run_reindex_command"
-    }
-}
-
-fn append_prepare_harness_warning_from_guidance(
-    warnings: &mut Vec<Value>,
-    warning_codes: &mut HashSet<String>,
-    guidance: &Value,
-    fallback_code: &str,
-    fallback_message: &str,
-    fallback_action: &str,
-    fallback_target: &str,
-) {
-    let code = guidance
-        .get("reason_code")
-        .and_then(|value| value.as_str())
-        .unwrap_or(fallback_code);
-    let message = guidance
-        .get("reason")
-        .and_then(|value| value.as_str())
-        .unwrap_or(fallback_message);
-    let recommended_action = guidance
-        .get("recommended_action")
-        .and_then(|value| value.as_str())
-        .unwrap_or(fallback_action);
-    let action_target = guidance
-        .get("action_target")
-        .and_then(|value| value.as_str())
-        .unwrap_or(fallback_target);
-    let mut extras = serde_json::Map::new();
-    for key in ["included_in", "recommended_profile"] {
-        if let Some(value) = guidance.get(key) {
-            extras.insert(key.to_owned(), value.clone());
-        }
-    }
-    if extras.is_empty() {
-        push_prepare_harness_warning(
-            warnings,
-            warning_codes,
-            code,
-            message,
-            action_target == "daemon" || code == "stale_daemon_binary",
-            recommended_action,
-            action_target,
-        );
-        return;
-    }
-    push_prepare_harness_warning_with_extras(
-        warnings,
-        warning_codes,
-        code,
-        message,
-        action_target == "daemon" || code == "stale_daemon_binary",
-        recommended_action,
-        action_target,
-        Value::Object(extras),
-    );
-}
-
-fn collect_prepare_harness_warnings(
-    capabilities_payload: &Value,
-    include_diagnostics_warning: bool,
-) -> Vec<Value> {
-    let mut warnings = Vec::new();
-    let mut warning_codes = HashSet::new();
-
-    if let Some(items) = capabilities_payload
-        .get("health_summary")
-        .and_then(|value| value.get("warnings"))
-        .and_then(|value| value.as_array())
-    {
-        for warning in items {
-            let code = warning
-                .get("code")
-                .and_then(|value| value.as_str())
-                .unwrap_or("health_warning");
-            let message = warning
-                .get("message")
-                .and_then(|value| value.as_str())
-                .unwrap_or("health warning");
-            let recommended_action = warning
-                .get("recommended_action")
-                .and_then(|value| value.as_str())
-                .unwrap_or("inspect_health_status");
-            let action_target = warning
-                .get("action_target")
-                .and_then(|value| value.as_str())
-                .unwrap_or("project");
-            push_prepare_harness_warning(
-                &mut warnings,
-                &mut warning_codes,
-                code,
-                message,
-                action_target == "daemon" || code == "stale_daemon_binary",
-                recommended_action,
-                action_target,
-            );
-        }
-    }
-
-    if let Some(guidance) = capabilities_payload
-        .get("semantic_search_guidance")
-        .filter(|value| {
-            !value
-                .get("available")
-                .and_then(|available| available.as_bool())
-                .unwrap_or(false)
-        })
-    {
-        append_prepare_harness_warning_from_guidance(
-            &mut warnings,
-            &mut warning_codes,
-            guidance,
-            "semantic_search_unavailable",
-            "semantic_search is unavailable",
-            "inspect_semantic_configuration",
-            "semantic_search",
-        );
-    }
-
-    if include_diagnostics_warning
-        && let Some(guidance) = capabilities_payload
-            .get("diagnostics_guidance")
-            .filter(|value| {
-                !value
-                    .get("available")
-                    .and_then(|available| available.as_bool())
-                    .unwrap_or(false)
-            })
-    {
-        append_prepare_harness_warning_from_guidance(
-            &mut warnings,
-            &mut warning_codes,
-            guidance,
-            "diagnostics_unavailable",
-            "diagnostics are unavailable",
-            "inspect_lsp_configuration",
-            "diagnostics",
-        );
-    }
-
-    warnings
-}
-
-fn index_stats_payload(stats: &codelens_engine::IndexStats) -> Value {
-    json!({
-        "indexed_files": stats.indexed_files,
-        "supported_files": stats.supported_files,
-        "stale_files": stats.stale_files,
-    })
-}
-
-fn prepare_harness_index_recovery(state: &AppState, arguments: &Value) -> Value {
-    let enabled = arguments
-        .get("auto_refresh_stale")
-        .and_then(|value| value.as_bool())
-        .unwrap_or(true);
-    let threshold = arguments
-        .get("auto_refresh_stale_threshold")
-        .and_then(|value| value.as_u64())
-        .map(|value| value as usize)
-        .unwrap_or(DEFAULT_AUTO_REFRESH_STALE_THRESHOLD);
-
-    let before = match state.symbol_index().stats() {
-        Ok(stats) => stats,
-        Err(error) => {
-            return json!({
-                "enabled": enabled,
-                "threshold": threshold,
-                "status": "unavailable",
-                "reason": "stats_unavailable",
-                "error": error.to_string(),
-            });
-        }
-    };
-
-    if !enabled {
-        return json!({
-            "enabled": false,
-            "threshold": threshold,
-            "status": "disabled",
-            "before": index_stats_payload(&before),
-        });
-    }
-
-    if before.stale_files == 0 {
-        return json!({
-            "enabled": true,
-            "threshold": threshold,
-            "status": "not_needed",
-            "before": index_stats_payload(&before),
-            "after": index_stats_payload(&before),
-        });
-    }
-
-    if before.stale_files > threshold {
-        return json!({
-            "enabled": true,
-            "threshold": threshold,
-            "status": "skipped",
-            "reason": "stale_threshold_exceeded",
-            "before": index_stats_payload(&before),
-        });
-    }
-
-    match state.symbol_index().refresh_all() {
-        Ok(after) => {
-            state.graph_cache().invalidate();
-            json!({
-                "enabled": true,
-                "threshold": threshold,
-                "status": "refreshed",
-                "reason": "stale_detected",
-                "before": index_stats_payload(&before),
-                "after": index_stats_payload(&after),
-            })
-        }
-        Err(error) => json!({
-            "enabled": true,
-            "threshold": threshold,
-            "status": "failed",
-            "reason": "refresh_failed",
-            "error": error.to_string(),
-            "before": index_stats_payload(&before),
-        }),
-    }
-}
-
-pub fn activate_project(state: &AppState, arguments: &serde_json::Value) -> ToolResult {
-    // If a project path is provided, switch the active project
-    let switched = if let Some(path) = arguments.get("project").and_then(|v| v.as_str()) {
-        match state.switch_project(path) {
-            Ok(name) => Some(name),
-            Err(e) => {
-                return Err(crate::error::CodeLensError::NotFound(format!(
-                    "failed to switch project: {e}"
-                )));
-            }
-        }
-    } else {
-        None
-    };
-
-    let project = state.project();
-    let project_name = project
-        .as_path()
-        .file_name()
-        .map(|n| n.to_string_lossy().to_string())
-        .unwrap_or_default();
-    let memories_dir = state.memories_dir();
-    let memory_count = list_memory_names(&memories_dir, None).len();
-    let watcher_running = state.watcher_running();
-    let frameworks = detect_frameworks(project.as_path());
-    let project_base_path = project.as_path().to_string_lossy().to_string();
-
-    // v1.5 Phase 2j MCP follow-up: auto-set `CODELENS_EMBED_HINT_AUTO_LANG`
-    // based on the project's dominant source language. Shared with the
-    // startup path (`main.rs`) so one-shot CLI and stdio MCP share the
-    // same detection + gating.
-    auto_set_embed_hint_lang(project.as_path());
-
-    // Auto-set role surface based on project size + client profile
-    let session = crate::session_context::SessionRequestContext::from_json(arguments);
-    let client = session
-        .client_name
-        .as_deref()
-        .map(|name| crate::client_profile::ClientProfile::detect(Some(name)))
-        .unwrap_or_else(|| state.client_profile());
-    let file_count = state
-        .symbol_index()
-        .stats()
-        .map(|s| s.indexed_files)
-        .unwrap_or(0);
-    // For Claude Code clients, keep Balanced preset (all tools accessible).
-    // Profile auto-selection only applies to Codex/generic clients.
-    let (auto_surface, auto_budget, auto_label) =
-        if matches!(client, crate::client_profile::ClientProfile::Claude) {
-            (
-                ToolSurface::Preset(ToolPreset::Balanced),
-                client.default_budget(),
-                "balanced",
-            )
-        } else if file_count < 50 {
-            (
-                ToolSurface::Profile(ToolProfile::BuilderMinimal),
-                default_budget_for_profile(ToolProfile::BuilderMinimal)
-                    .max(client.default_budget()),
-                "builder-minimal",
-            )
-        } else if file_count > 500 {
-            (
-                ToolSurface::Profile(ToolProfile::ReviewerGraph),
-                default_budget_for_profile(ToolProfile::ReviewerGraph).max(client.default_budget()),
-                "reviewer-graph",
-            )
-        } else {
-            (
-                ToolSurface::Profile(ToolProfile::PlannerReadonly),
-                default_budget_for_profile(ToolProfile::PlannerReadonly)
-                    .max(client.default_budget()),
-                "planner-readonly",
-            )
-        };
-    #[cfg(feature = "http")]
-    if state.should_route_to_session(&session) {
-        state.set_session_surface_and_budget(&session.session_id, auto_surface, auto_budget);
-        state.bind_project_to_session(&session.session_id, &project_base_path);
-    } else {
-        state.set_surface(auto_surface);
-        state.set_token_budget(auto_budget);
-    }
-    #[cfg(not(feature = "http"))]
-    {
-        state.set_surface(auto_surface);
-        state.set_token_budget(auto_budget);
-    }
-
-    #[cfg(feature = "semantic")]
-    let embedding_ready = state.embedding_ref().is_some();
-    #[cfg(not(feature = "semantic"))]
-    let embedding_ready = false;
-
-    Ok((
-        json!({
-            "activated": true,
-            "switched": switched.is_some(),
-            "project_name": project_name,
-            "project_base_path": project_base_path,
-            "backend_id": "rust-core",
-            "memory_count": memory_count,
-            "serena_memories_dir": memories_dir.to_string_lossy(),
-            "file_watcher": watcher_running,
-            "frameworks": frameworks,
-            "auto_surface": auto_label,
-            "auto_budget": auto_budget,
-            "indexed_files": file_count,
-            "embedding_ready": embedding_ready
-        }),
-        success_meta(BackendKind::Session, 1.0),
-    ))
-}
-
-pub fn prepare_harness_session(state: &AppState, arguments: &serde_json::Value) -> ToolResult {
-    // Issue #224: previously this rejected callers that supplied both
-    // `preset` and `profile`, but the schema documented both as
-    // independent optional fields and routing docs frequently pair
-    // them ("profile=builder-minimal" + "preset=minimal"). The
-    // rejection burned a bootstrap round trip.
-    //
-    // New contract: `profile` wins (it carries the semantic role that
-    // determines tool surface + executor bias), `preset` is dropped
-    // when both are supplied, and the dropped value is surfaced via
-    // `surface_resolution` + a non-blocking warning so the caller
-    // can adjust on the next call.
-    let requested_profile = arguments
-        .get("profile")
-        .and_then(|v| v.as_str())
-        .map(str::to_owned);
-    let requested_preset = arguments
-        .get("preset")
-        .and_then(|v| v.as_str())
-        .map(str::to_owned);
-    let explicit_project_request = arguments.get("project").and_then(|v| v.as_str()).is_some();
-    let preset_dropped_for_profile = requested_profile.is_some() && requested_preset.is_some();
-    // Drop `preset` for the downstream `activate_project` call so
-    // the existing single-knob path runs unchanged. Working on a
-    // clone keeps the caller's original arguments intact for logging
-    // / response echo.
-    let adjusted_arguments_owner: Option<serde_json::Value> =
-        preset_dropped_for_profile.then(|| {
-            let mut cloned = arguments.clone();
-            if let Some(map) = cloned.as_object_mut() {
-                map.remove("preset");
-            }
-            cloned
-        });
-    let effective_arguments: &serde_json::Value =
-        adjusted_arguments_owner.as_ref().unwrap_or(arguments);
-
-    // Preserve existing surface when caller does not explicitly request a
-    // profile/preset change. `activate_project` auto-selects a surface based
-    // on file count, which overwrites a harness-chosen surface on every
-    // bootstrap call. We snapshot before activation and restore unless the
-    // user provided a new profile/preset.
-    let prior_surface = *state.surface();
-    let explicit_surface_request = requested_profile.is_some() || requested_preset.is_some();
-
-    let (activate_payload, _) = activate_project(state, effective_arguments)?;
-
-    // Restore surface if the caller did not explicitly ask for a new one.
-    if !explicit_surface_request {
-        state.set_surface(prior_surface);
-    }
-
-    // Apply effort_level if provided (before preset/profile for budget calculation)
-    if let Some(effort_str) = arguments.get("effort_level").and_then(|v| v.as_str()) {
-        let level = match effort_str {
-            "low" => crate::client_profile::EffortLevel::Low,
-            "medium" => crate::client_profile::EffortLevel::Medium,
-            _ => crate::client_profile::EffortLevel::High,
-        };
-        state.set_effort_level(level);
-    }
-
-    if arguments.get("profile").and_then(|v| v.as_str()).is_some() {
-        crate::tools::session::set_profile(state, arguments)?;
-    } else if arguments.get("preset").and_then(|v| v.as_str()).is_some() {
-        crate::tools::session::set_preset(state, arguments)?;
-    } else if let Some(budget) = arguments
-        .get("token_budget")
-        .and_then(|v| v.as_u64())
-        .map(|v| v as usize)
-    {
-        #[cfg(feature = "http")]
-        {
-            let session = crate::session_context::SessionRequestContext::from_json(arguments);
-            if state.should_route_to_session(&session) {
-                state.set_session_surface_and_budget(
-                    &session.session_id,
-                    state.execution_surface(&session),
-                    budget,
-                );
-            } else {
-                state.set_token_budget(budget);
-            }
-        }
-        #[cfg(not(feature = "http"))]
-        {
-            state.set_token_budget(budget);
-        }
-    }
-
-    let index_recovery = prepare_harness_index_recovery(state, arguments);
-
-    let detail = arguments
-        .get("detail")
-        .and_then(|v| v.as_str())
-        .unwrap_or("full");
-    let request = ResourceRequestContext::from_request("codelens://tools/list", Some(arguments));
-    let session = request.session.clone();
-    let active_surface = state.execution_surface(&session);
-    let token_budget = state.execution_token_budget(&session);
-    let visible = build_visible_tool_context(state, &request);
-    let surface_generation =
-        crate::tool_schema_generation::surface_generation_payload(&visible.tools);
-    let current_tool_schema_fingerprint =
-        crate::tool_schema_generation::tool_schema_fingerprint(&visible.tools);
-    let reported_client_tool_schema_fingerprint =
-        client_tool_schema_fingerprint(arguments).map(str::to_owned);
-    let requested_host_context = arguments
-        .get("host_context")
-        .and_then(|value| value.as_str());
-    let requested_task_overlay = arguments
-        .get("task_overlay")
-        .and_then(|value| value.as_str());
-    let host_context = requested_host_context.and_then(HostContext::from_str);
-    let task_overlay = requested_task_overlay.and_then(TaskOverlay::from_str);
-    let overlay_plan = compile_surface_overlay(active_surface, host_context, task_overlay);
-
-    let config_payload = if detail == "full" {
-        let (payload, _) = crate::tools::filesystem::get_current_config(state, arguments)?;
-        payload
-    } else {
-        json!({
-            "runtime": "rust-core",
-            "project_root": state.project().as_path().display().to_string(),
-            "surface": active_surface.as_label(),
-            "token_budget": token_budget,
-            "tool_count": crate::tool_defs::visible_tools(active_surface).len(),
-            "client_profile": request.client_profile.as_str(),
-        })
-    };
-    let capabilities_arguments = match arguments.get("file_path").and_then(|v| v.as_str()) {
-        Some(file_path) => json!({ "file_path": file_path }),
-        None => json!({}),
-    };
-    let (capabilities_payload, _) =
-        crate::tools::session::get_capabilities(state, &capabilities_arguments)?;
-    let health_summary = capabilities_payload
-        .get("health_summary")
-        .cloned()
-        .unwrap_or_else(|| json!({"status": "ok", "warning_count": 0, "warnings": []}));
-    let warnings = collect_prepare_harness_warnings(
-        &capabilities_payload,
-        arguments
-            .get("file_path")
-            .and_then(|value| value.as_str())
-            .is_some(),
-    );
-    let warnings = {
-        let mut warnings = warnings;
-        let mut warning_codes = warnings
-            .iter()
-            .filter_map(|warning| {
-                warning
-                    .get("code")
-                    .and_then(|value| value.as_str())
-                    .map(str::to_owned)
-            })
-            .collect::<HashSet<_>>();
-        if let Some(client_fingerprint) = reported_client_tool_schema_fingerprint.as_deref()
-            && client_fingerprint != current_tool_schema_fingerprint
-        {
-            push_prepare_harness_warning_with_extras(
-                &mut warnings,
-                &mut warning_codes,
-                "tool_schema_cache_stale",
-                "client-reported tool schema fingerprint does not match the active CodeLens tool surface; refresh tools/list or reconnect before trusting cached tool input schemas",
-                true,
-                crate::tool_schema_generation::TOOL_SCHEMA_REFRESH_ACTION,
-                "tool_schema_cache",
-                json!({
-                    "client_tool_schema_fingerprint": client_fingerprint,
-                    "server_tool_schema_fingerprint": current_tool_schema_fingerprint,
-                    "schema_version": crate::surface_manifest::SURFACE_MANIFEST_SCHEMA_VERSION,
-                    "refresh": {
-                        "method": "tools/list",
-                        "params": { "full": true },
-                        "fallback": "reconnect_mcp_server"
-                    },
-                }),
-            );
-        }
-        match index_recovery
-            .get("status")
-            .and_then(|value| value.as_str())
-            .unwrap_or("unknown")
-        {
-            "failed" => {
-                let stale_files = index_recovery
-                    .get("before")
-                    .and_then(|before| before.get("stale_files"))
-                    .and_then(|value| value.as_u64())
-                    .unwrap_or(0);
-                push_prepare_harness_warning_with_extras(
-                    &mut warnings,
-                    &mut warning_codes,
-                    "index_refresh_failed",
-                    index_recovery
-                        .get("error")
-                        .and_then(|value| value.as_str())
-                        .unwrap_or("failed to refresh stale index during bootstrap"),
-                    false,
-                    refresh_symbol_index_recommended_action_for_surface(active_surface),
-                    "symbol_index",
-                    json!({
-                        "remediation": refresh_symbol_index_remediation_for_surface(
-                            active_surface,
-                            RefreshSymbolIndexRemediation::Force
-                        ),
-                        "stale_files": stale_files,
-                    }),
-                )
-            }
-            "skipped" => {
-                let stale_files = index_recovery
-                    .get("before")
-                    .and_then(|before| before.get("stale_files"))
-                    .and_then(|value| value.as_u64())
-                    .unwrap_or(0);
-                let threshold = index_recovery
-                    .get("threshold")
-                    .and_then(|value| value.as_u64())
-                    .unwrap_or(DEFAULT_AUTO_REFRESH_STALE_THRESHOLD as u64);
-                push_prepare_harness_warning_with_extras(
-                    &mut warnings,
-                    &mut warning_codes,
-                    "index_refresh_skipped",
-                    "stale index detected but auto-refresh threshold was exceeded",
-                    false,
-                    refresh_symbol_index_recommended_action_for_surface(active_surface),
-                    "symbol_index",
-                    json!({
-                        "remediation": refresh_symbol_index_remediation_for_surface(
-                            active_surface,
-                            RefreshSymbolIndexRemediation::StaleOnly
-                        ),
-                        "auto_refresh_threshold": {
-                            "max_stale_files": threshold,
-                            "current_stale_files": stale_files,
-                        },
-                    }),
-                )
-            }
-            _ => {}
-        }
-        if requested_host_context.is_some() && host_context.is_none() {
-            push_prepare_harness_warning(
-                &mut warnings,
-                &mut warning_codes,
-                "unknown_host_context",
-                "prepare_harness_session received an unknown host_context hint and fell back to surface-default routing",
-                false,
-                "use_documented_host_context",
-                "bootstrap_routing",
-            );
-        }
-        if requested_task_overlay.is_some() && task_overlay.is_none() {
-            push_prepare_harness_warning(
-                &mut warnings,
-                &mut warning_codes,
-                "unknown_task_overlay",
-                "prepare_harness_session received an unknown task_overlay hint and fell back to surface-default routing",
-                false,
-                "use_documented_task_overlay",
-                "bootstrap_routing",
-            );
-        }
-        // Issue #186: detect when the active project resolved to an
-        // anonymized agent identifier (e.g. `agent-a110134bd9c6e7440`)
-        // instead of the daemon's CLI startup root. This happens when
-        // a session-bound switch lands on an internal Claude/Codex
-        // workspace path where the directory basename is itself a
-        // hash. The active project is then "shared on-disk index"
-        // material from a sibling daemon's view but the answer key is
-        // a different tree, surfacing as `indexed_files: 0` +
-        // false-positive `no_supported_files`. Surface the mismatch
-        // so the caller can re-issue with `project=<real path>`.
-        let active_project_name_for_check = activate_payload
-            .get("project_name")
-            .and_then(|value| value.as_str())
-            .unwrap_or_default();
-        let active_project_is_anonymized =
-            is_anonymized_agent_project_name(active_project_name_for_check);
-        if active_project_is_anonymized {
-            let daemon_default = state.default_project_scope();
-            push_prepare_harness_warning_with_extras(
-                &mut warnings,
-                &mut warning_codes,
-                "anonymized_project_name_detected",
-                &format!(
-                    "active project resolved to anonymized agent identifier `{active_project_name_for_check}`; this usually means a session-bound switch landed on an internal harness workspace and the on-disk index for the daemon's CLI default project is not loaded. Re-issue prepare_harness_session with `project=<absolute repo path>` to pin the intended project.",
-                ),
-                false,
-                "activate_explicit_project",
-                "active_project",
-                json!({
-                    "anonymized_project_name": active_project_name_for_check,
-                    "daemon_default_project_root": daemon_default,
-                    "remediation": {
-                        "method": "tool_call",
-                        "tool": "prepare_harness_session",
-                        "args": { "project": daemon_default },
-                    },
-                }),
-            );
-        }
-        if !explicit_project_request && !active_project_is_anonymized {
-            let active_project_root = state.current_project_scope();
-            let daemon_default_project_root = state.default_project_scope();
-            if active_project_root != daemon_default_project_root {
-                let suggested_project = daemon_default_project_root.clone();
-                push_prepare_harness_warning_with_extras(
-                    &mut warnings,
-                    &mut warning_codes,
-                    "active_project_differs_from_daemon_default",
-                    "active CodeLens project differs from the daemon default project. If this is not the workspace you intend to inspect, re-issue prepare_harness_session or activate_project with an absolute project path; do not fall back to native tools solely because the active project is stale.",
-                    false,
-                    "verify_or_activate_explicit_project",
-                    "active_project",
-                    json!({
-                        "active_project_root": active_project_root,
-                        "daemon_default_project_root": daemon_default_project_root,
-                        "native_fallback_recommended": false,
-                        "remediation": {
-                            "tool": "prepare_harness_session",
-                            "args": {
-                                "project": suggested_project.clone(),
-                                "detail": "compact"
-                            },
-                            "alternative_tool": "activate_project",
-                            "alternative_args": {
-                                "project": suggested_project
-                            }
-                        }
-                    }),
-                );
-            }
-        }
-        // Issue #224: when caller supplied both `profile` and `preset`,
-        // surface a non-blocking warning naming the dropped value so the
-        // next call can drop the redundant arg explicitly. Profile wins
-        // because it carries the semantic role + executor bias.
-        if preset_dropped_for_profile {
-            let profile_str = requested_profile.as_deref().unwrap_or("?");
-            let preset_str = requested_preset.as_deref().unwrap_or("?");
-            push_prepare_harness_warning_with_extras(
-                &mut warnings,
-                &mut warning_codes,
-                "preset_dropped_for_profile",
-                &format!(
-                    "both `profile` and `preset` supplied; using profile=`{profile_str}` and dropping preset=`{preset_str}` (profile wins)",
-                ),
-                false,
-                "drop_redundant_argument",
-                "preset",
-                json!({
-                    "winner_field": "profile",
-                    "winner_value": requested_profile,
-                    "dropped_field": "preset",
-                    "dropped_value": requested_preset,
-                }),
-            );
-        }
-        warnings
-    };
-
-    let visible_tool_names = visible
-        .tools
-        .iter()
-        .map(|tool| tool.name.to_owned())
-        .collect::<Vec<_>>();
-    let requested_entrypoints = arguments
-        .get("preferred_entrypoints")
-        .and_then(|value| value.as_array())
-        .map(|items| {
-            items
-                .iter()
-                .filter_map(|item| item.as_str())
-                .map(ToOwned::to_owned)
-                .collect::<Vec<_>>()
-        })
-        .unwrap_or_default();
-    let overlay_preferred_entrypoints = overlay_plan
-        .preferred_entrypoints
-        .iter()
-        .map(|tool| (*tool).to_owned())
-        .collect::<Vec<_>>();
-    let preferred_entrypoints_source = if !requested_entrypoints.is_empty() {
-        "provided"
-    } else if !overlay_preferred_entrypoints.is_empty() {
-        "overlay"
-    } else {
-        "surface_default"
-    };
-    let preferred_entrypoint_requests = if !requested_entrypoints.is_empty() {
-        tool_name_requests(requested_entrypoints)
-    } else if !overlay_preferred_entrypoints.is_empty() {
-        tool_name_requests(overlay_preferred_entrypoints)
-    } else {
-        tool_name_requests(
-            preferred_bootstrap_tools(active_surface)
-                .unwrap_or(&[])
-                .iter()
-                .map(|tool| (*tool).to_owned())
-                .collect::<Vec<_>>(),
-        )
-    };
-    let preferred_entrypoints = preferred_entrypoint_requests
-        .iter()
-        .map(|request| request.tool.clone())
-        .collect::<Vec<_>>();
-    let preferred_entrypoints_visible = preferred_entrypoints
-        .iter()
-        .filter(|tool| visible_tool_names.iter().any(|name| name == *tool))
-        .cloned()
-        .collect::<Vec<_>>();
-    let overlay_preferred_entrypoints_visible = overlay_plan
-        .preferred_entrypoints
-        .iter()
-        .filter(|tool| visible_tool_names.iter().any(|name| name == *tool))
-        .map(|tool| (*tool).to_owned())
-        .collect::<Vec<_>>();
-    let overlay_emphasized_tools = overlay_plan
-        .emphasized_tools
-        .iter()
-        .map(|tool| (*tool).to_owned())
-        .collect::<Vec<_>>();
-    let overlay_emphasized_tools_visible = overlay_plan
-        .emphasized_tools
-        .iter()
-        .filter(|tool| visible_tool_names.iter().any(|name| name == *tool))
-        .map(|tool| (*tool).to_owned())
-        .collect::<Vec<_>>();
-    let overlay_avoid_tools = overlay_plan
-        .avoid_tools
-        .iter()
-        .map(|tool| (*tool).to_owned())
-        .collect::<Vec<_>>();
-    let overlay_avoid_tools_visible = overlay_plan
-        .avoid_tools
-        .iter()
-        .filter(|tool| visible_tool_names.iter().any(|name| name == *tool))
-        .map(|tool| (*tool).to_owned())
-        .collect::<Vec<_>>();
-    let preferred_entrypoints_with_executors = preferred_entrypoints_visible
-        .iter()
-        .map(|tool| {
-            json!({
-                "tool": tool,
-                "preferred_executor": crate::tool_defs::tool_preferred_executor_label(tool),
-            })
-        })
-        .collect::<Vec<_>>();
-    let preferred_entrypoints_omitted = tool_request_omissions(
-        &preferred_entrypoint_requests,
-        &preferred_entrypoints_visible,
-        active_surface,
-        visible.deferred_loading_active,
-    );
-    let recommended_entrypoint = preferred_entrypoints_visible.first().cloned();
-    let recommended_entrypoint_preferred_executor = recommended_entrypoint
-        .as_deref()
-        .map(crate::tool_defs::tool_preferred_executor_label);
-    let mut visible_executor_counts = std::collections::BTreeMap::new();
-    for tool in &visible.tools {
-        *visible_executor_counts
-            .entry(crate::tool_defs::tool_preferred_executor_label(tool.name).to_owned())
-            .or_insert(0usize) += 1;
-    }
-
-    let result = if detail == "full" {
-        json!({
-            "activated": true,
-            "project": activate_payload,
-            "active_surface": active_surface.as_label(),
-            "token_budget": token_budget,
-            "surface_generation": surface_generation,
-            "config": config_payload,
-            "index_recovery": index_recovery,
-            "capabilities": capabilities_payload,
-            "health_summary": health_summary,
-            "warnings": warnings,
-            "overlay": {
-                "applied": overlay_plan.applied(),
-                "host_context": overlay_plan.host_context.map(|value| value.as_str()),
-                "task_overlay": overlay_plan.task_overlay.map(|value| value.as_str()),
-                "preferred_executor_bias": overlay_plan.preferred_executor_bias,
-                "preferred_entrypoints": overlay_plan.preferred_entrypoints,
-                "preferred_entrypoints_visible": overlay_preferred_entrypoints_visible,
-                "emphasized_tools": overlay_emphasized_tools,
-                "emphasized_tools_visible": overlay_emphasized_tools_visible,
-                "avoid_tools": overlay_avoid_tools,
-                "avoid_tools_visible": overlay_avoid_tools_visible,
-                "routing_notes": overlay_plan.routing_notes,
-            },
-            "http_session": build_http_session_payload(state, &request),
-            "visible_tools": {
-                "tool_count": visible.tools.len(),
-                "tool_count_total": visible.total_tool_count,
-                "tool_names": visible_tool_names,
-                "preferred_executors": visible_executor_counts,
-                "all_namespaces": visible.all_namespaces,
-                "all_tiers": visible.all_tiers,
-                "preferred_namespaces": visible.preferred_namespaces,
-                "preferred_tiers": visible.preferred_tiers,
-                "loaded_namespaces": visible.loaded_namespaces,
-                "loaded_tiers": visible.loaded_tiers,
-                "effective_namespaces": visible.effective_namespaces,
-                "effective_tiers": visible.effective_tiers,
-                "selected_namespace": visible.selected_namespace,
-                "selected_tier": visible.selected_tier,
-                "deferred_loading_active": visible.deferred_loading_active,
-                "full_tool_exposure": visible.full_tool_exposure,
-            },
-            "routing": {
-                "preferred_entrypoints": preferred_entrypoints,
-                "preferred_entrypoints_source": preferred_entrypoints_source,
-                "preferred_entrypoints_visible": preferred_entrypoints_visible,
-                "preferred_entrypoints_omitted": preferred_entrypoints_omitted,
-                "preferred_entrypoints_with_executors": preferred_entrypoints_with_executors,
-                "recommended_entrypoint": recommended_entrypoint,
-                "recommended_entrypoint_preferred_executor": recommended_entrypoint_preferred_executor,
-            },
-            "harness": {
-                "effort_level": state.effort_level().as_str(),
-                "compression_offset": state.effort_level().compression_threshold_offset(),
-                "meta_max_result_size": true,
-                "rapid_burst_detection": true,
-                "schema_pre_validation": true,
-                "doom_loop_threshold": 3,
-                "preflight_ttl_seconds": state.preflight_ttl_seconds(),
-            }
-        })
-    } else {
-        let project_name = activate_payload
-            .get("project_name")
-            .and_then(|v| v.as_str())
-            .unwrap_or_default();
-        let indexed_files = activate_payload
-            .get("indexed_files")
-            .and_then(|v| v.as_u64())
-            .unwrap_or(0);
-        let capabilities_available = capabilities_payload
-            .get("available")
-            .cloned()
-            .unwrap_or_else(|| json!([]));
-        let tool_count = visible.tools.len();
-        // Issue #199-B-1: compact mode trims `tool_names` to the first 5
-        // and `preferred_entrypoints_visible` to whatever the routing
-        // layer surfaces, but it never tells the caller *how much* was
-        // dropped. The full-detail response carries `*_omitted_count`
-        // markers next to every trimmed array; compact must do the
-        // same so callers can budget their next call instead of
-        // re-issuing `detail=full` just to learn the surface size.
-        const COMPACT_TOOL_NAMES_LIMIT: usize = 5;
-        let first_five_tools: Vec<_> = visible_tool_names
-            .iter()
-            .take(COMPACT_TOOL_NAMES_LIMIT)
-            .cloned()
-            .collect();
-        let tool_names_omitted_count = visible_tool_names
-            .len()
-            .saturating_sub(first_five_tools.len());
-        let preferred_entrypoints_visible_omitted_count = preferred_entrypoints
-            .len()
-            .saturating_sub(preferred_entrypoints_visible.len());
-        let health_status = health_summary
-            .get("status")
-            .and_then(|v| v.as_str())
-            .unwrap_or("ok");
-        json!({
-            "activated": true,
-            "project": {
-                "project_name": project_name,
-                "indexed_files": indexed_files,
-            },
-            "capabilities": {
-                "available": capabilities_available,
-            },
-            "surface_generation": surface_generation,
-            "visible_tools": {
-                "tool_count": tool_count,
-                "tool_names": first_five_tools,
-                "tool_names_omitted_count": tool_names_omitted_count,
-            },
-            "health_summary": {
-                "status": health_status,
-            },
-            "warnings": warnings,
-            "routing": {
-                "recommended_entrypoint": recommended_entrypoint,
-                "preferred_entrypoints_visible": preferred_entrypoints_visible,
-                "preferred_entrypoints_omitted": preferred_entrypoints_omitted,
-                "preferred_entrypoints_visible_omitted_count":
-                    preferred_entrypoints_visible_omitted_count,
-            },
-        })
-    };
-
-    Ok((result, success_meta(BackendKind::Session, 1.0)))
-}
-
-/// v1.5 Phase 2j MCP follow-up: auto-detect and export the dominant source
-/// language for the given project so the engine's `auto_hint_should_enable`
-/// gate can consult `language_supports_nl_stack` on the next embedding call.
-///
-/// Applied at two entry points:
-///   1. Startup in `main.rs` — covers one-shot CLI (`--cmd`) and stdio MCP.
-///   2. `activate_project` — covers MCP-driven project switches.
-///
-/// Only fires when:
-///   (1) auto mode is explicitly enabled via `CODELENS_EMBED_HINT_AUTO=1`
-///       (default-OFF policy held — no automatic behaviour change),
-///   (2) the user has not already set `CODELENS_EMBED_HINT_AUTO_LANG`
-///       themselves (explicit > auto, same rule as the three per-gate
-///       env vars).
-///
-/// The detection walk is capped at 16k files inside
-/// `compute_dominant_language` so even large monorepos pay a bounded cost.
-/// When the walk yields no confident answer (fewer than 3 source files, or
-/// no known-extension files at all), we leave the env var unset and the
-/// engine falls through to the conservative default (stack OFF).
-pub fn auto_set_embed_hint_lang(project_path: &std::path::Path) {
-    // v1.6.0 flip (§8.14): default-ON semantics. Unset env means "auto
-    // mode ON", explicit `CODELENS_EMBED_HINT_AUTO=0`/`false`/`no`/`off`
-    // is the opt-out. Must stay in lock-step with the engine's
-    // `auto_hint_mode_enabled()` in `crates/codelens-engine/src/embedding/mod.rs`.
-    let auto_hint_gate_enabled = std::env::var("CODELENS_EMBED_HINT_AUTO")
-        .ok()
-        .map(|v| {
-            let lowered = v.trim().to_ascii_lowercase();
-            match lowered.as_str() {
-                "1" | "true" | "yes" | "on" => true,
-                "0" | "false" | "no" | "off" => false,
-                _ => true, // unknown value → fall through to default-on
-            }
-        })
-        .unwrap_or(true);
-    let user_forced_lang = std::env::var("CODELENS_EMBED_HINT_AUTO_LANG").is_ok();
-    if !auto_hint_gate_enabled || user_forced_lang {
-        return;
-    }
-    let Some(lang) = compute_dominant_language(project_path) else {
-        return;
-    };
-    // Export to the process environment so the engine's gate functions
-    // (`nl_tokens_enabled`, `api_calls_enabled`, `sparse_weighting_enabled`)
-    // read it on the next call. Process-scoped — startup sets it once, and
-    // `activate_project` re-writes it on project switch (handled via
-    // `user_forced_lang` short-circuit: if we switch projects we'd have to
-    // clear the var first, which is an acceptable follow-up limitation).
-    //
-    // SAFETY: `set_var` is unsafe on modern Rust because env-var mutation
-    // is not thread-safe. Both call sites (startup main + single-threaded
-    // MCP request handler) run before the engine has spawned worker
-    // threads that read env, so the concurrent-read hazard does not apply.
-    unsafe {
-        std::env::set_var("CODELENS_EMBED_HINT_AUTO_LANG", lang);
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn extras_warning_merges_object_keys_and_dedupes_by_code() {
-        let mut warnings: Vec<Value> = Vec::new();
-        let mut codes: HashSet<String> = HashSet::new();
-
-        push_prepare_harness_warning_with_extras(
-            &mut warnings,
-            &mut codes,
-            "index_refresh_skipped",
-            "stale index detected but auto-refresh threshold was exceeded",
-            false,
-            "refresh_symbol_index",
-            "symbol_index",
-            json!({
-                "remediation": {
-                    "method": "tool_call",
-                    "tool": "refresh_symbol_index",
-                    "args": { "scope": "stale_only" },
-                    "alternative_command": "codelens reindex --stale-only",
-                },
-                "auto_refresh_threshold": {
-                    "max_stale_files": 32,
-                    "current_stale_files": 47,
-                },
-            }),
-        );
-
-        // Same code is dropped on the second call — no duplicate warnings.
-        push_prepare_harness_warning_with_extras(
-            &mut warnings,
-            &mut codes,
-            "index_refresh_skipped",
-            "second emission must be ignored",
-            false,
-            "refresh_symbol_index",
-            "symbol_index",
-            json!({ "remediation": { "tool": "ignored" } }),
-        );
-
-        assert_eq!(warnings.len(), 1, "duplicate code must be deduped");
-        let warning = warnings[0].as_object().expect("warning is object");
-        assert_eq!(warning["code"], json!("index_refresh_skipped"));
-        assert_eq!(warning["recommended_action"], json!("refresh_symbol_index"));
-        assert_eq!(warning["action_target"], json!("symbol_index"));
-        assert_eq!(warning["restart_recommended"], json!(false));
-        let remediation = warning["remediation"].as_object().expect("remediation");
-        assert_eq!(remediation["tool"], json!("refresh_symbol_index"));
-        assert_eq!(remediation["args"]["scope"], json!("stale_only"));
-        assert_eq!(
-            remediation["alternative_command"],
-            json!("codelens reindex --stale-only")
-        );
-        let threshold = warning["auto_refresh_threshold"]
-            .as_object()
-            .expect("threshold");
-        assert_eq!(threshold["max_stale_files"], json!(32));
-        assert_eq!(threshold["current_stale_files"], json!(47));
-    }
-
-    #[test]
-    fn extras_warning_ignores_non_object_extras() {
-        let mut warnings: Vec<Value> = Vec::new();
-        let mut codes: HashSet<String> = HashSet::new();
-
-        push_prepare_harness_warning_with_extras(
-            &mut warnings,
-            &mut codes,
-            "index_refresh_failed",
-            "failed to refresh stale index during bootstrap",
-            false,
-            "refresh_symbol_index",
-            "symbol_index",
-            // Non-object extras must be dropped silently so the warning shape
-            // stays consistent — callers should not have to validate `extras`
-            // construction.
-            json!("not-an-object"),
-        );
-
-        assert_eq!(warnings.len(), 1);
-        let warning = warnings[0].as_object().expect("warning is object");
-        assert_eq!(warning["code"], json!("index_refresh_failed"));
-        // No spurious key inserted from the string extras.
-        assert!(warning.get("remediation").is_none());
-        assert!(warning.get("auto_refresh_threshold").is_none());
-    }
-
-    #[test]
-    fn refresh_index_remediation_marks_hidden_tool_uncallable() {
-        let remediation = refresh_symbol_index_remediation_for_surface(
-            ToolSurface::Profile(ToolProfile::ReviewerGraph),
-            RefreshSymbolIndexRemediation::StaleOnly,
-        );
-
-        assert_eq!(remediation["method"], json!("shell"));
-        assert_eq!(
-            remediation["command"],
-            json!("codelens reindex --stale-only")
-        );
-        assert_eq!(
-            remediation["alternative_command"],
-            json!("codelens reindex --stale-only")
-        );
-        assert_eq!(
-            remediation["tool_call"]["tool"],
-            json!("refresh_symbol_index")
-        );
-        assert_eq!(
-            remediation["tool_call"]["args"]["scope"],
-            json!("stale_only")
-        );
-        assert_eq!(remediation["tool_call"]["callable"], json!(false));
-        assert_eq!(
-            remediation["tool_call"]["reason"],
-            json!("not_in_active_surface")
-        );
-        assert_eq!(remediation["tool_call"]["surface"], json!("reviewer-graph"));
-    }
-
-    #[test]
-    fn refresh_index_remediation_keeps_visible_tool_call_primary() {
-        let remediation = refresh_symbol_index_remediation_for_surface(
-            ToolSurface::Profile(ToolProfile::BuilderMinimal),
-            RefreshSymbolIndexRemediation::StaleOnly,
-        );
-
-        assert_eq!(remediation["method"], json!("tool_call"));
-        assert_eq!(remediation["tool"], json!("refresh_symbol_index"));
-        assert_eq!(remediation["args"]["scope"], json!("stale_only"));
-        assert_eq!(remediation["callable"], json!(true));
-        assert_eq!(
-            remediation["alternative_command"],
-            json!("codelens reindex --stale-only")
-        );
-    }
-
-    #[test]
-    fn refresh_index_recommended_action_matches_surface_callability() {
-        assert_eq!(
-            refresh_symbol_index_recommended_action_for_surface(ToolSurface::Profile(
-                ToolProfile::BuilderMinimal
-            )),
-            "refresh_symbol_index"
-        );
-        assert_eq!(
-            refresh_symbol_index_recommended_action_for_surface(ToolSurface::Profile(
-                ToolProfile::ReviewerGraph
-            )),
-            "run_reindex_command"
-        );
-    }
-
-    #[test]
-    fn guidance_warning_preserves_semantic_surface_hints() {
-        let mut warnings: Vec<Value> = Vec::new();
-        let mut codes: HashSet<String> = HashSet::new();
-
-        append_prepare_harness_warning_from_guidance(
-            &mut warnings,
-            &mut codes,
-            &json!({
-                "reason_code": "semantic_not_in_active_surface",
-                "reason": "not in active surface",
-                "recommended_action": "switch_tool_surface",
-                "action_target": "tool_surface",
-                "included_in": ["planner-readonly", "builder-minimal"],
-                "recommended_profile": "planner-readonly",
-            }),
-            "semantic_search_unavailable",
-            "semantic_search is unavailable",
-            "inspect_semantic_configuration",
-            "semantic_search",
-        );
-
-        assert_eq!(warnings.len(), 1);
-        let warning = warnings[0].as_object().expect("warning is object");
-        assert_eq!(warning["code"], json!("semantic_not_in_active_surface"));
-        assert_eq!(warning["recommended_action"], json!("switch_tool_surface"));
-        assert_eq!(warning["action_target"], json!("tool_surface"));
-        assert_eq!(warning["recommended_profile"], json!("planner-readonly"));
-        assert_eq!(
-            warning["included_in"],
-            json!(["planner-readonly", "builder-minimal"])
-        );
-    }
-
-    /// Issue #186 detector: a `agent-<hash>` directory basename is the
-    /// telltale sign that the active project is an internal Claude/
-    /// Codex workspace, not the daemon's CLI startup root.
-    #[test]
-    fn anonymized_agent_project_name_detected_for_hash_pattern() {
-        // Real-world cases from the dogfood report.
-        assert!(is_anonymized_agent_project_name("agent-a110134bd9c6e7440"));
-        assert!(is_anonymized_agent_project_name(
-            "agent-0123456789abcdefABCDEF"
-        ));
-        // Minimum-length boundary (12 chars after prefix).
-        assert!(is_anonymized_agent_project_name("agent-aaaaaaaaaaaa"));
-    }
-
-    /// Real project directories that happen to start with `agent`
-    /// must not trip the detector. Only hash-shaped suffixes count.
-    #[test]
-    fn anonymized_agent_project_name_rejects_real_project_names() {
-        assert!(!is_anonymized_agent_project_name("agent-server")); // dash inside not hex
-        assert!(!is_anonymized_agent_project_name("agent-cli")); // too short
-        assert!(!is_anonymized_agent_project_name("codelens-mcp-plugin"));
-        assert!(!is_anonymized_agent_project_name("rg-family"));
-        assert!(!is_anonymized_agent_project_name("agent-orchestrator")); // dash inside
-        // A sub-12 hash-ish suffix is also rejected (too short to be the
-        // anonymizer the harness uses).
-        assert!(!is_anonymized_agent_project_name("agent-abc123"));
-    }
-
-    /// Edge cases: empty string, missing prefix, only the prefix.
-    #[test]
-    fn anonymized_agent_project_name_handles_edge_cases() {
-        assert!(!is_anonymized_agent_project_name(""));
-        assert!(!is_anonymized_agent_project_name("agent-"));
-        assert!(!is_anonymized_agent_project_name("foo-agent-aaaaaaaaaaaa"));
-        assert!(!is_anonymized_agent_project_name("AGENT-aaaaaaaaaaaa")); // case sensitive
-    }
-}
