+++ crates/codelens-mcp/src/tools/session/project_ops/mod.rs
@@ -1,0 +1,13 @@
+pub mod activate;
+pub mod embed_hint;
+pub mod prep_recovery;
+pub mod prep_warnings;
+pub mod prepare_harness;
+pub mod util;
+
+#[cfg(test)]
+mod tests;
+
+pub use activate::activate_project;
+pub use embed_hint::auto_set_embed_hint_lang;
+pub use prepare_harness::prepare_harness_session;
